samples: bluetooth: power benchmark samples
Adds a test sample applications to collect measurements for
various adveritsing and connection scenarios on nRF SoCs.

The central is a tester application, that should be used to
connect to a DK that measurements are collected from.

The peripheral is a DUT application, that should be run on
a DK to take emasurements.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>