samples: bluetooth: Peripheral HR using SDC peripeheral variant for radio FW
The commit provides modified peripehral_hr_coded sample that
uses SDC peripehral variant for nRF5420 Radio core FW.

The sample is bare minimum that allows to establish a connection with it.
There is no HR service. There is no coded PHY.

This is a dirty change to existing sample to provide example based on
current sdk-nrf main and make it possible to a proof of concept that
was done some time ago to run the radio core FW from local RAM.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>